requirements.txt
pandas
numpy
mysql-connector-python
pymysql
scikit-learn
python-dotenv